// src/lib/api/models/getApiFieldDefinitionsEntityType.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * RealHub API
 * Multi-tenant Real Estate Platform API
 * OpenAPI spec version: v1
 */

export type GetApiFieldDefinitionsEntityType = typeof GetApiFieldDefinitionsEntityType[keyof typeof GetApiFieldDefinitionsEntityType];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const GetApiFieldDefinitionsEntityType = {
  PROPERTY: 'PROPERTY',
  CUSTOMER_NEED: 'CUSTOMER_NEED',
  LEAD: 'LEAD',
  DEAL: 'DEAL',
  OWNER_PROFILE: 'OWNER_PROFILE',
} as const;